requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
asyncpg>=0.12.0
click
pandas>=0.23.4
xlrd==1.2.0
requests
gitpython
pygsheets==2.0.0
importlib
parsedatetime
coc.py
lru-dict